fix(#46): Adjust sample to use proper scenario hierarchy
