+++ simulator-samples/sample-rest/src/main/java/com/consol/citrus/simulator/sample/scenario/DefaultGreetingScenario.java
@@ -1,0 +1,12 @@
+package com.consol.citrus.simulator.sample.scenario;
+
+/**
+ * @author Christoph Deppisch
+ */
+public class DefaultGreetingScenario extends AbstractGreetingScenario {
+
+    @Override
+    String getGreetingMessage() {
+        return "What's up!";
+    }
+}
