+++ simulator-samples/sample-rest/src/main/java/com/consol/citrus/simulator/sample/scenario/HowdyScenario.java
@@ -4,12 +4,12 @@ import com.consol.citrus.simulator.scenario.Scenario;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-@Scenario("Hello")
-@RequestMapping(value = "/services/rest/simulator/hello", method = RequestMethod.POST)
-public class FriendlyHelloScenario extends HelloScenario{
+@Scenario("Howdy")
+@RequestMapping(value = "/services/rest/simulator/howdy", method = RequestMethod.POST)
+public class HowdyScenario extends DefaultGreetingScenario {
 
     @Override
-    String getMessage() {
-        return "there";
+    String getGreetingMessage() {
+        return "Howdy partner!";
     }
 }
